apps/ai-engine/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-genai
pypdf
python-docx
httpx